change int to float in total_carton_weight_in_kg delivery note

+++ bbt_bpm/custom_script/delivery_note/delivery_note.py
@@ -212,7 +212,7 @@ def set_items(doc):
 			boxes.total_carton_weight_in_kg = boxes.total_carton_weight_in_kg 
 
 		box.append(boxes.carton_qty)
-		total_craton_weight.append(int(boxes.total_carton_weight_in_kg))
+		total_craton_weight.append(float(boxes.total_carton_weight_in_kg))
 	doc.total_no_of_boxes = sum(box)
 	doc.total_craton_weight = sum(total_craton_weight)
